Refactor qtv scraper.
