+++ internal/app/app.go
@@ -12,7 +12,7 @@ import (
 	"github.com/gofiber/fiber/v2/middleware/cors"
 	"github.com/gofiber/fiber/v2/middleware/favicon"
 	"github.com/gofiber/fiber/v2/middleware/recover"
-	"github.com/vikpe/qw-hub-api/pkg/demoscraper/qtv"
+	"github.com/vikpe/qw-hub-api/pkg/qtvscraper"
 	"github.com/vikpe/qw-hub-api/pkg/serverscraper"
 	"github.com/vikpe/qw-hub-api/pkg/twitch"
 )
@@ -21,7 +21,7 @@ type Config struct {
 	Port           int                  `json:"port"`
 	Servers        serverscraper.Config `json:"servers"`
 	Streamers      twitch.StreamerIndex `json:"streamers"`
-	QtvDemoSources []qtv.Server         `json:"qtv_demo_sources"`
+	QtvDemoSources []qtvscraper.Server  `json:"qtv_demo_sources"`
 }
 
 func New() *fiber.App {
