+++ pkg/qtvscraper/server_test.go
@@ -1,24 +1,24 @@
-package qtv_test
+package qtvscraper_test
 
 import (
 	"testing"
 	"time"
 
-	"github.com/go-playground/assert/v2"
-	"github.com/vikpe/qw-hub-api/pkg/demoscraper/qtv"
+	"github.com/stretchr/testify/assert"
 	"github.com/vikpe/qw-hub-api/pkg/qdemo"
+	"github.com/vikpe/qw-hub-api/pkg/qtvscraper"
 )
 
 func TestQtvServer(t *testing.T) {
-	server := qtv.Server{Address: "qw.foppa.dk:28000"}
+	server := qtvscraper.Server{Address: "qw.foppa.dk:28000"}
 	assert.Equal(t, "http://qw.foppa.dk:28000/dl/demos/foo.mvd", server.DemoDownloadUrl("foo.mvd"))
 	assert.Equal(t, "file:foo.mvd@qw.foppa.dk:28000", server.DemoQtvplayUrl("foo.mvd"))
 }
 
 func TestQtvHostedDemo(t *testing.T) {
-	demo := qtv.Demo{
+	demo := qtvscraper.Demo{
 		Filename: qdemo.Filename("duel_holy_vs_si7h[bravado]261022-2255.mvd"),
-		Server: qtv.Server{
+		Server: qtvscraper.Server{
 			Address:        "troopers.fi:28000",
 			DemoDateFormat: "dmy",
 		},
